requirements.txt
pydantic==2.12.5
pyyaml==6.0.3
pytest==9.0.2